% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmHeterosced.R
\name{lmmHeterosced}
\alias{lmmHeterosced}
\title{Linear mixed model with heteroscedasticity}
\usage{
lmmHeterosced(formula, data, heterosced_formula = ~1, REML = TRUE)
}
\arguments{
\item{formula}{model formula lmer style}

\item{data}{a data frame with the data}

\item{heterosced_formula}{an optional heteroscedasticity formula}

\item{REML}{if restricted maximum likelihood should be used}
}
\value{
\code{lmmHeterosced} returns a list with elements:
Fixef (fixed effects estimates),  
Heterosced (parameters of the heteroscedasticity function, a linear function of the log residual variance),  
Ranef_log_Var (random effect variances on log),  
Ranef (random effect),
Residuals,
data (the original data),
error_var_matrix (variance matrix of the estimation error)
fit (nlminb output),
obj (MakeADFun output),
optTime (Optimization time)
}
\description{
\code{lmmHeterosced} Linear mixed model with the log error variance as linear function of data.
}
\examples{
# See the vignette 'lmmHeterosced'.
}
\author{
Geir H. Bolstad
}
