% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmHeterosced.R
\name{plot.lmmHeterosced}
\alias{plot.lmmHeterosced}
\title{Plot of lmmHeterosced object}
\usage{
\method{plot}{lmmHeterosced}(
  x,
  x_axis,
  col_data = "grey",
  col_expectation = "black",
  col_sd = "blue",
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{'lmmHeterosced'}.}

\item{x_axis}{The name of the explanatory variable that should be plotted.}

\item{col_data}{The colour of the datapoints.}

\item{col_expectation}{The colour of the regression line showing the expectation}

\item{col_sd}{The colour of the regression lines showing +/- one standard deviation}

\item{xlab}{as in \code{\link{plot}}.}

\item{ylab}{as in \code{\link{plot}}.}

\item{...}{Additional arguments passed to \code{\link{plot}}.}

\item{col}{as in \code{\link{plot}}.}
}
\value{
\code{plot} returns a plot of the lmmHeterosced regression.
}
\description{
\code{plot} method for class \code{'lmmHeterosced'}.
}
\details{
Plots the regression fitted by the \code{\link{lmmHeterosced}}
  function.
}
\examples{
# See the vignette 'lmmHeterosced'.
}
\author{
Geir H. Bolstad
}
