% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmHeterosced.R
\name{plot_abs_y}
\alias{plot_abs_y}
\title{Plot of absolute y values}
\usage{
plot_abs_y(
  mod,
  x,
  xlab = NULL,
  ylab = NULL,
  col = "grey",
  line_col = "black",
  cex.legend = 1,
  ...
)
}
\arguments{
\item{mod}{An object of class \code{'lmmHeterosced'}.}

\item{x}{The name of the explanatory variable that should be plotted.}

\item{xlab}{as in \code{\link{plot}}.}

\item{ylab}{as in \code{\link{plot}}.}

\item{col}{as in \code{\link{plot}}.}

\item{line_col}{The colour of the regression line.}

\item{cex.legend}{A character expansion factor relative to current par("cex")
for the printed parameters.}

\item{...}{Additional arguments passed to \code{\link{plot}}.}
}
\value{
A plot.
}
\description{
\code{plot_abs_y} plots the expected change in the absolute values of y from an object of class \code{'lmmHeterosced'}.
}
\examples{
# See the vignette 'Analyzing rates of evolution'.
}
\author{
Geir H. Bolstad
}
