% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmHeterosced.R
\name{lmmHeterosced}
\alias{lmmHeterosced}
\title{Linear mixed model with heteroscedastisity}
\usage{
lmmHeterosced(formula, data, heterosced_formula = ~1)
}
\arguments{
\item{formula}{model formula lmer style}

\item{data}{a data frame with the data}

\item{heterosced_formula}{an optional heteroscadisity formula}
}
\value{
\code{lmmHeterosced} a list with elements:
Fixef (fixed effects estaimtes),
Heterosced (parameters of the heteroscedasticity function, a linear function of the log residual variance),
Ranef_log_Var (random effect variances on log),
Ranef (random effect),
fit (nlminb output),
obj (MakeADFun ouptut),
optTime (Optimization time)
}
\description{
\code{lmmHeterosced} Linear mixed model with the log error variance as linear function of data.
}
\examples{

}
\author{
Geir H. Bolstad
}

